+++ package.json
@@ -14,7 +14,7 @@
   "author": "TotalTechGeek",
   "license": "BSD-2-Clause",
   "dependencies": {
-    "ajv": "^6.12.3",
+    "ajv": "^8.10.0",
     "express": "^4.17.1",
     "fs-extra": "^8.1.0",
     "json-power-query": "^1.0.1",
