+++ src/controller/Controller.ts
@@ -1,0 +1,6 @@
+import {Action} from 'src/controller/Action';
+
+export type Controller = {
+  actions: Array<Action>
+}
+
