+++ src/controller/RestRoute.ts
@@ -1,0 +1,6 @@
+export type RestRoute<TRoute extends string, TRouteParams extends string> = {
+  type: 'rest',
+  pattern: TRoute;
+  params: Array<TRouteParams>;
+}
+
