+++ src/controller/coerce/controller.ts
@@ -1,0 +1,5 @@
+import {Controller} from '../Controller';
+
+export function controller(controller: Controller): Controller {
+  throw 'TODO'
+}
