+++ src/controller/coerce/index.ts
@@ -1,0 +1,4 @@
+export * from './rest';
+export * from './extract';
+export * from './action';
+export * from './controller';
