+++ src/extract/Extract.ts
@@ -1,0 +1,6 @@
+
+import * as extractors from './extractors';
+
+export const Extract = Object.freeze({
+  from: Object.freeze(extractors)
+});
