+++ src/extract/RequestExtractor.ts
@@ -1,0 +1,4 @@
+export type RequestExtractor<TParamName extends string, TOutput> = {
+  paramName?: TParamName,
+  apply: () => TOutput
+}
