+++ src/extract/SimpleExtraction.ts
@@ -1,0 +1,5 @@
+export type SimpleExtraction<TParamName extends string,
+  TExtract extends object> = {
+  paramName?: TParamName;
+  apply: () => TExtract;
+}
