+++ src/extract/extractors/body.ts
@@ -1,0 +1,5 @@
+import {SimpleExtraction} from 'src/extract/SimpleExtraction';
+
+export function body<TExtract extends object>(): SimpleExtraction<never, TExtract> {
+  throw 'todo';
+}
