+++ src/types/Awaitable.ts
@@ -1,0 +1,1 @@
+export type Awaitable<T> = T | PromiseLike<T>;
