+++ src/types/HttpMethod.ts
@@ -1,0 +1,2 @@
+export type HttpMethod = 'delete' | 'get' | 'patch' | 'post' | 'put';
+
