+++ tsconfig.json
@@ -3,7 +3,8 @@
     "outDir": "./out",
     "allowJs": true,
     "target": "es2020",
-    "baseUrl": "./src"
+    "baseUrl": "./",
+    "moduleResolution": "node"
   },
   "include": [
     "./src/**/*"
